# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Paw Time Platform API
  version: 0.1.0
  description: ワーカーアプリと採用企業アプリを接続する共通API。
servers:
  - url: http://localhost:8787
paths:
...
  /v1/business/applications/{applicationId}/decision:
    post:
      operationId: decideApplication
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 採用または不採用を記録
  /v1/business/shifts/{shiftId}/attendance:
    post:
      operationId: recordAttendance
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: 追記型の出退勤イベントを記録
  /v1/business/shifts/{shiftId}/attendance-summary:
    get:
      operationId: getAttendanceSummary
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 予定時刻と実績から算出した遅刻・出退勤情報
...